# packages/contracts/openapi/v1/openapi.api.json
{
  "components": {
    "schemas": {
...
      "Artifact": {
        "additionalProperties": false,
        "description": "图表、数据、代码、日志、论文等交付物的元数据。内容本体在对象存储中内容寻址，服务端负责重新计算并核验 sha256。",
        "properties": {
          "created_at": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "id": {
            "pattern": "^art_[0-9a-f]{32}$",
            "title": "Id",
            "type": "string"
          },
          "inputs": {
            "description": "上游 Artifact 血缘；失效传播沿此边计算。",
            "items": {
              "pattern": "^art_[0-9a-f]{32}$",
              "type": "string"
            },
            "title": "Inputs",
            "type": "array"
          },
          "kind": {
            "$ref": "#/components/schemas/omm_contracts__v1__artifact__Kind"
          },
          "media_type": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Media Type",
            "type": "string"
          },
          "producer_step_id": {
            "anyOf": [
              {
                "pattern": "^step_[0-9a-f]{32}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Producer Step Id"
          },
          "project_id": {
            "pattern": "^proj_[0-9a-f]{32}$",
            "title": "Project Id",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "pattern": "^run_[0-9a-f]{32}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "sha256": {
            "$ref": "#/components/schemas/Sha256"
          },
          "size_bytes": {
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/omm_contracts__v1__artifact__Status"
          },
          "uri": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "kind",
          "uri",
          "sha256",
          "size_bytes",
          "media_type",
          "inputs",
          "status",
          "created_at"
        ],
        "title": "Artifact",
        "type": "object"
      },
      "ArtifactList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Artifact"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "ArtifactList",
        "type": "object"
      },
...
      "ArtifactText": {
        "description": "附件正文抽取结果。\n\n``status`` 分五档：ready 完整抽出、partial 触顶截断、empty 文件正常但没有\n文字、unsupported 缺少可选依赖或格式不支持、failed 文件损坏或抽取出错。\n后三档也是正常响应（200）——调用方需要的是原因，而不是一个错误码。",
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "characters": {
            "title": "Characters",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "segments": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segments"
          },
          "status": {
            "enum": [
              "ready",
              "partial",
              "empty",
              "unsupported",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "artifact_id",
          "name",
          "media_type",
          "status",
          "engine",
          "characters",
          "text"
        ],
        "title": "ArtifactText",
        "type": "object"
      },
...
      "Body_upload_project_artifact_api_v1_projects__project_id__artifacts_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "kind": {
            "default": "other",
            "title": "Kind",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_project_artifact_api_v1_projects__project_id__artifacts_post",
        "type": "object"
      },
...
      "CreateProjectInput": {
        "properties": {
          "competition_policy": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competition Policy"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Mode"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "workspace_uri": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Uri"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateProjectInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Mode": {
        "description": "产品模式，见规划文档 §1.2。",
        "enum": [
          "learning",
          "collaboration",
          "auto_experiment",
          "review",
          "organization"
        ],
        "title": "Mode",
        "type": "string"
      },
...
      "Project": {
        "additionalProperties": false,
        "description": "一个持续存在的建模项目。领域对象事实来源为 PostgreSQL，本契约描述 API 对外表示。",
        "properties": {
          "competition_policy": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "CompetitionPolicyProfile 引用 ID；MVP 允许为空。",
            "title": "Competition Policy"
          },
          "created_at": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "$ref": "#/components/schemas/ProjectId"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "owner": {
            "description": "所有者标识。MVP 单用户阶段固定为 local-dev，接入认证后为用户 ID。",
            "maxLength": 200,
            "minLength": 1,
            "title": "Owner",
            "type": "string"
          },
          "updated_at": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "workspace_uri": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "工作区根 URI（对象存储前缀或本地路径）。",
            "title": "Workspace Uri"
          }
        },
        "required": [
          "id",
          "name",
          "owner",
          "mode",
          "created_at",
          "updated_at"
        ],
        "title": "Project",
        "type": "object"
      },
      "ProjectId": {
        "pattern": "^proj_[0-9a-f]{32}$",
        "title": "ProjectId",
        "type": "string"
      },
      "ProjectList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "ProjectList",
        "type": "object"
      },
...
      "Sha256": {
        "pattern": "^[a-f0-9]{64}$",
        "title": "Sha256",
        "type": "string"
      },
...
      "Timestamp": {
        "description": "UTC ISO-8601，统一以 Z 结尾。",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,6})?Z$",
        "title": "Timestamp",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "omm_contracts__v1__artifact__Kind": {
        "enum": [
          "dataset",
          "code",
          "figure",
          "table",
          "log",
          "report",
          "paper",
          "model",
          "other"
        ],
        "title": "Kind",
        "type": "string"
      },
      "omm_contracts__v1__artifact__Status": {
        "enum": [
          "PENDING",
          "READY",
          "STALE",
          "DELETED"
        ],
        "title": "Status",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "OpenMathModel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/artifacts/{artifact_id}/download": {
      "get": {
        "operationId": "download_artifact_api_v1_artifacts__artifact_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}/text": {
      "get": {
        "description": "读取附件正文；首次访问时现场抽取并缓存。\n\n抽取放在读取时而不是上传时：上传路径要对用户即时响应，而几十兆的 PDF 抽一遍\n要好几秒。产物内容寻址、字节不可变，因此抽一次可以一直复用；``refresh=true``\n用于服务端补装了 OCR 或解析依赖之后重跑。",
        "operationId": "read_artifact_text_api_v1_artifacts__artifact_id__text_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactText"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Artifact Text",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/projects": {
      "get": {
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/artifacts": {
      "get": {
        "operationId": "list_project_artifacts_api_v1_projects__project_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Project Artifacts",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "description": "上传产物：服务端计算 sha256 并内容寻址存储（主规划 §19 对象存储/哈希校验）。",
        "operationId": "upload_project_artifact_api_v1_projects__project_id__artifacts_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_project_artifact_api_v1_projects__project_id__artifacts_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artifact"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Project Artifact",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}